Add common component for daily/spotter data charts
